[OPS-6907] I cannot make it work unless it is a 100% separate plugin, with separate config. Sorry.
