+++ src/main/resources/com/flowdock/jenkins/FlowdockBuildStart/config.jelly
@@ -1,0 +1,12 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+
+    <f:entry title="Flow API token(s)" field="flowToken">
+        <f:textbox />
+    </f:entry>
+
+    <f:entry title="Notification tags" field="notificationTags">
+        <f:textbox />
+    </f:entry>
+
+</j:jelly>
